Use trait for utilities method for accessing data. So the entity class size will not increase a lot when we add additional field
