+++ src/Entity/Player.php
@@ -1,10 +1,11 @@
 <?php namespace GameScan\WoW\Entity;
 
 use GameScan\WoW\WowApiRequest;
+use GameScan\WoW\Entity\Player\Base as PlayerBase;
 
 class Player extends Base
 {
-
+    use PlayerBase;
     protected $realmName;
     protected $characterName;
 
@@ -24,59 +25,5 @@ class Player extends Base
         return "wow/character/" . $this->realmName . "/" . $this->characterName;
     }
 
-    public function getLastModified()
-    {
-        return $this->getEntityInformations()->lastInformation;
-    }
 
-    public function name()
-    {
-        return $this->getEntityInformations()->name;
-    }
-
-    public function realm()
-    {
-        return $this->getEntityInformations()->realm;
-    }
-
-    public function playerClass()
-    {
-        return $this->getEntityInformations()->class;
-    }
-
-    public function race()
-    {
-        return $this->getEntityInformations()->race;
-    }
-
-    public function gender()
-    {
-        return $this->getEntityInformations()->gender;
-    }
-
-    public function level()
-    {
-        return $this->getEntityInformations()->level;
-    }
-
-    public function achievementPoints()
-    {
-        return $this->getEntityInformations()->achievementPoints;
-    }
-
-
-    public function thumbnail()
-    {
-        return $this->getEntityInformations()->thumbnail;
-    }
-
-    public function calcClass()
-    {
-        return $this->getEntityInformations()->calcClass;
-    }
-
-    public function totalHonorableKills()
-    {
-        return $this->getEntityInformations()->totalHonorableKills;
-    }
 }
